Created Python package genv
* Separated genv.py into three modules
